refactor: validator 패키지로 Validor 클래스 이동
