+++ src/main/java/racingcar/validator/CarNameValidator.java
@@ -1,11 +1,11 @@
-package racingcar.util;
+package racingcar.validator;
 
 public class CarNameValidator {
 
     private static final int CAR_NAME_MAX_LENGTH = 5;
 
     public static boolean isWithinMaxLength(String carName) {
-        
+
         return carName.length() > CAR_NAME_MAX_LENGTH;
     }
