+++ src/main/java/racingcar/validator/TryCountValidator.java
@@ -1,4 +1,4 @@
-package racingcar.util;
+package racingcar.validator;
 
 import java.util.regex.Pattern;
 
@@ -7,7 +7,7 @@ public class TryCountValidator {
     private static final Pattern POSITIVE_INTEGER = Pattern.compile("^[1-9]\\d*$");
 
     public static boolean isPositiveNumber(String tryCount) {
-        
+
         return POSITIVE_INTEGER.matcher(tryCount).matches();
     }
 }
